requirements.txt
arabic-reshaper==2.1.3
asgiref==3.3.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.4
cssselect2==0.4.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.10
django-heroku==0.3.1
django-mssql-backend==2.8.1
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
heroku==0.1.4
html5lib==1.1
idna==3.2
Pillow==8.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.20
pyodbc==4.0.30
PyPDF2==1.26.0
pyphen==0.11.0
python-bidi==0.4.2
python-dateutil==1.5
pytz==2021.1
reportlab==3.5.68
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.6
WeasyPrint==52.5
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlwt==1.3.0
xlwt-fix==1.3.1
